+++ addon/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-07-01 21:35 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-07-02 13:17 UTC] — do not edit by hand.*
 
 # Changelog